add join

+++ src/com/company/interfaces/components/ScoresPanel.java
@@ -16,18 +16,12 @@ public class ScoresPanel extends JPanel {
         this.user = user;
         System.out.println("Из scores Panel : " + user.toString());
         ResultSet resultSet = ScoreController.getAllScores(user);
-//        Object[][] array = new Object[resultSet.size()][];
-//        for (int i = 0; i < score.getResultList().size(); i++) {
-//            array[i] = score.getResultList().get(i);
-//        }
-
-//        JTable table = new JTable(resultSet, new Object[] {"id", "title", "sum", "userId"});
         DefaultTableModel tableModel = new DefaultTableModel(new Object[]{"N", "Cчет", "Баланс"}, 0);
         try {
             int number = 1;
             // Проходим по всем строкам resultSet, добавляя каждую строку в модель таблицы
             while (resultSet.next()) {
-                Object[] row = {number, resultSet.getString("title"), resultSet.getInt("sum")};
+                Object[] row = {number, resultSet.getString("title"), resultSet.getDouble("sum")};
                 tableModel.addRow(row);
                 number++;
             }
